new tabs implementation

+++ app/src/main/java/com/phlox/tvwebbrowser/model/WebTabState.kt
@@ -5,25 +5,27 @@ import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.os.Bundle
 import android.webkit.WebChromeClient
-import com.phlox.tvwebbrowser.activity.main.MainActivity
+import com.phlox.tvwebbrowser.TVBro
 
 import com.phlox.tvwebbrowser.activity.main.view.WebViewEx
 import com.phlox.tvwebbrowser.utils.Utils
-import org.json.JSONArray
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.launch
 import org.json.JSONException
 import org.json.JSONObject
 
 import java.io.ByteArrayOutputStream
 import java.io.File
 import java.io.FileOutputStream
-import java.util.ArrayList
+import java.nio.charset.Charset
 
 /**
  * Created by PDT on 24.08.2016.
  *
  * Class to store state of tab with webView
  */
-data class WebTabState(var currentOriginalUrl: String? = null, var currentTitle: String? = null,
+data class WebTabState(var currentOriginalUrl: String? = null, var currentTitle: String = "",
                        var selected: Boolean = false, var thumbnailHash: String? = null,
                        var faviconHash: String? = null, var thumbnail: Bitmap? = null,
                        var favicon: Bitmap? = null) {
@@ -45,12 +47,6 @@ data class WebTabState(var currentOriginalUrl: String? = null, var currentTitle:
             selected = json.getBoolean("selected")
             if (json.has("thumbnail")) {
                 thumbnailHash = json.getString("thumbnail")
-                val thumbnailFile = File(context.cacheDir.absolutePath +
-                        File.separator + WebTabState.TAB_THUMBNAILS_DIR +
-                        File.separator + thumbnailHash)
-                if (thumbnailFile.exists()) {
-                    thumbnail = BitmapFactory.decodeFile(thumbnailFile.absolutePath)
-                }
             }
             if (json.has("favicon")) {
                 faviconHash = json.getString("favicon")
@@ -67,49 +63,17 @@ data class WebTabState(var currentOriginalUrl: String? = null, var currentTitle:
 
     }
 
-    fun toJson(context: Context, storeFiles: Boolean): JSONObject {
+    fun toJson(): JSONObject {
         val store = JSONObject()
         try {
             store.put("url", currentOriginalUrl)
             store.put("title", currentTitle)
             store.put("selected", selected)
-            if (storeFiles) {
-                val tabsThumbsDir = File(context.cacheDir.absolutePath + File.separator + WebTabState.TAB_THUMBNAILS_DIR)
-                if (tabsThumbsDir.exists() || tabsThumbsDir.mkdir()) {
-                    if (thumbnail != null) {
-                        try {
-                            val baos = ByteArrayOutputStream()
-                            thumbnail!!.compress(Bitmap.CompressFormat.PNG, 100, baos)
-                            val bitmapBytes = baos.toByteArray()
-                            val hash = Utils.MD5_Hash(bitmapBytes)
-                            if (hash != null && hash != thumbnailHash) {
-                                if (thumbnailHash != null) {
-                                    removeThumbnailFile(context)
-                                    thumbnailHash = null
-                                }
-                                val file = File(tabsThumbsDir.absolutePath + File.separator + hash)
-                                var fos: FileOutputStream? = null
-                                try {
-                                    fos = FileOutputStream(file)
-                                    fos.write(bitmapBytes)
-                                } catch (e: Exception) {
-                                    e.printStackTrace()
-                                } finally {
-                                    fos?.close()
-                                }
-
-                                thumbnailHash = hash
-                                store.put("thumbnail", thumbnailHash)
-                            }
-                        } catch (e: Exception) {
-                            e.printStackTrace()
-                        }
-                    }
-                }
-
-                if (faviconHash != null) {
-                    store.put("favicon", faviconHash)
-                }
+            if (thumbnailHash != null) {
+                store.put("thumbnail", thumbnailHash)
+            }
+            if (faviconHash != null) {
+                store.put("favicon", faviconHash)
             }
         } catch (e: JSONException) {
             e.printStackTrace()
@@ -118,17 +82,56 @@ data class WebTabState(var currentOriginalUrl: String? = null, var currentTitle:
         return store
     }
 
-    fun removeFiles(context: Context) {
-        if (thumbnailHash != null) {
-            removeThumbnailFile(context)
+    private fun saveThumbnail(context: Context, scope: CoroutineScope) {
+        val thumbnail = this.thumbnail
+        val thumbnailHash = this.thumbnailHash
+        val url = currentOriginalUrl
+        if (thumbnail == null || url == null) return
+        scope.launch(Dispatchers.IO) {
+            synchronized(this@WebTabState) {
+                val tabsThumbsDir = File(context.cacheDir.absolutePath + File.separator + TAB_THUMBNAILS_DIR)
+                if (tabsThumbsDir.exists() || tabsThumbsDir.mkdir()) {
+                    try {
+                        val hash = Utils.MD5_Hash(url.toByteArray(Charset.defaultCharset()))
+                        if (hash != null && hash != thumbnailHash) {
+                            if (thumbnailHash != null) {
+                                removeThumbnailFile()
+                            }
+                            val file = File(getThumbnailPath(hash))
+                            var fos: FileOutputStream? = null
+                            try {
+                                fos = FileOutputStream(file)
+                                thumbnail.compress(Bitmap.CompressFormat.PNG, 100, fos)
+                            } catch (e: Exception) {
+                                e.printStackTrace()
+                            } finally {
+                                fos?.close()
+                            }
+
+                            this@WebTabState.thumbnailHash = hash
+                        }
+                    } catch (e: Exception) {
+                        e.printStackTrace()
+                    }
+                }
+            }
         }
     }
 
-    private fun removeThumbnailFile(context: Context) {
-        val thumbnailFile = File(context.cacheDir.absolutePath +
-                File.separator + WebTabState.TAB_THUMBNAILS_DIR +
-                File.separator + thumbnailHash)
+    private fun getThumbnailPath(hash: String) =
+        TVBro.instance.cacheDir.absolutePath + File.separator + TAB_THUMBNAILS_DIR + File.separator + hash + ".png"
+
+    fun removeFiles() {
+        if (thumbnailHash != null) {
+            removeThumbnailFile()
+        }
+    }
+
+    private fun removeThumbnailFile() {
+        if (thumbnailHash == null) return
+        val thumbnailFile = File(getThumbnailPath(thumbnailHash!!))
         thumbnailFile.delete()
+        thumbnailHash = null
     }
 
     fun restoreWebView() {
@@ -183,5 +186,32 @@ data class WebTabState(var currentOriginalUrl: String? = null, var currentTitle:
                 }
             }
         }
+    }
+
+    fun updateThumbnail(context: Context, thumbnail: Bitmap, scope: CoroutineScope) {
+        this.thumbnail = thumbnail
+        val url = currentOriginalUrl
+        if (url != null) {
+            var hash = Utils.MD5_Hash(url.toByteArray(Charset.defaultCharset()))
+            if (hash != null) {
+                hash += hashCode()//to make thumbnails from different tabs unique even with same url
+                if (hash != thumbnailHash) {
+                    saveThumbnail(context, scope)
+                }
+            }
+        }
+    }
+
+    fun loadThumbnail(): Bitmap? {
+        val hash = thumbnailHash ?: return null
+        val thumbnailFile = File(getThumbnailPath(hash))
+        if (thumbnailFile.exists()) {
+            thumbnail = BitmapFactory.decodeFile(thumbnailFile.absolutePath,
+                    BitmapFactory.Options().apply { this.inMutable = true })
+            return thumbnail
+        } else {
+            thumbnailHash = null
+        }
+        return null
     }
 }
